user login

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,23 @@ class TasksController extends Controller
      */
     public function index()
     {
-        /// タスク一覧を取得
-        $tasks = Task::all();
+         $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
+        
     }
 
     /**
@@ -52,11 +62,15 @@ class TasksController extends Controller
             'status' => 'required|max:10', 
             'content' => 'required|max:255',
         ]);
-       // タスクを作成
+        
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        
         $task = new Task;
+        $task->user_id = $request->user()->id;
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
+    
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -68,15 +82,22 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
-    {
-       // idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
+    public function show()
+    {   
+        // idの値でユーザを検索して取得
+        $user = User::findOrFail($id);
 
-        // タスク詳細ビューでそれを表示
+        // 関係するモデルの件数をロード
+        $user->loadRelationshipCounts();
+
+        // ユーザの投稿一覧を作成日時の降順で取得
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+        // ユーザ詳細ビューでそれらを表示
         return view('tasks.show', [
-            'task' => $task,
+            'tasks' => $tasks,
         ]);
+        
     }
 
     /**
@@ -129,10 +150,13 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-         // idの値でタスクを検索して取得
+         // idの値でタスクを検索して取得W
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
+        
+         // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
